cleanup warnings with ILP64
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@5004 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ control/xerbla.cpp
@@ -32,5 +32,5 @@ extern "C"
 void magma_xerbla(const char *srname , magma_int_t info)
 {
     fprintf( stderr, "On entry to %s, parameter %d had an illegal value\n",
-             srname, info );
+             srname, (int) info );
 }
