+++ testing/testing_veclib.cpp
@@ -126,8 +126,8 @@ float test( magma_int_t m, magma_int_t n )
         dnorm_one = ddot_(  &m, dA, &ione, dA, &ione );
         snorm_fro = snrm2_( &m, sA, &ione );
         dnorm_fro = dnrm2_( &m, dA, &ione );
-        printf( "m %d, sdot %12.8f, snrm2 %12.8f\n", m, snorm_one, snorm_fro );
-        printf( "m %d, ddot %12.8f, dnrm2 %12.8f\n", m, dnorm_one, dnorm_fro );
+        printf( "m %d, sdot %12.8f, snrm2 %12.8f\n", (int) m, snorm_one, snorm_fro );
+        printf( "m %d, ddot %12.8f, dnrm2 %12.8f\n", (int) m, dnorm_one, dnorm_fro );
         error = fabs(snorm_one - dnorm_one) / dnorm_one;
         status |= ! (error < 1e-6);
     }
@@ -149,10 +149,10 @@ float test( magma_int_t m, magma_int_t n )
         dnorm_fro = dlange_( "fro", &m, &n, dA, &lda, dwork );
         
         printf( "m %d, n %d, slange norm one %12.8f,  inf %12.8f,  max %12.8f,  fro %12.8f\n",
-                m, n, snorm_one, snorm_inf, snorm_max, snorm_fro );
+                (int) m, (int) n, snorm_one, snorm_inf, snorm_max, snorm_fro );
         
         printf( "m %d, n %d, dlange norm one %12.8f,  inf %12.8f,  max %12.8f,  fro %12.8f\n",
-                m, n, dnorm_one, dnorm_inf, dnorm_max, dnorm_fro );
+                (int) m, (int) n, dnorm_one, dnorm_inf, dnorm_max, dnorm_fro );
         error = fabs(snorm_one - dnorm_one) / dnorm_one;
         status |= ! (error < 1e-6);
     }
@@ -174,10 +174,10 @@ float test( magma_int_t m, magma_int_t n )
         dnorm_fro = dlansy_( "fro", "up", &n, dA, &lda, dwork );
         
         printf( "m %d, n %d, slansy norm one %12.8f,  inf %12.8f,  max %12.8f,  fro %12.8f\n",
-                m, n, snorm_one, snorm_inf, snorm_max, snorm_fro );
+                (int) m, (int) n, snorm_one, snorm_inf, snorm_max, snorm_fro );
         
         printf( "m %d, n %d, dlansy norm one %12.8f,  inf %12.8f,  max %12.8f,  fro %12.8f\n",
-                m, n, dnorm_one, dnorm_inf, dnorm_max, dnorm_fro );
+                (int) m, (int) n, dnorm_one, dnorm_inf, dnorm_max, dnorm_fro );
         error = fabs(snorm_one - dnorm_one) / dnorm_one;
         status |= ! (error < 1e-6);
     }
